+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2014-03-03
+* Fix depth length issue with long source i.e. github
+
+
 ## 2014-02-26
 * Move npmls script to node package
 * Update regex to allow linked packages